+++ app/src/main/java/com/example/graduationproject/MainActivity.java
@@ -1,127 +1,44 @@
 package com.example.graduationproject;
 
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.core.app.ActivityCompat;
-
-import android.Manifest;
-import android.content.pm.PackageManager;
+import android.content.Intent;
 import android.os.Bundle;
-import android.util.Log;
-import android.util.SparseArray;
-import android.view.SurfaceHolder;
-import android.view.SurfaceView;
-import android.widget.TextView;
+import android.view.View;
+import android.widget.Button;
 
-import com.google.android.gms.vision.CameraSource;
-import com.google.android.gms.vision.Detector;
-import com.google.android.gms.vision.text.TextBlock;
-import com.google.android.gms.vision.text.TextRecognizer;
+import androidx.appcompat.app.AppCompatActivity;
 
-import java.io.IOException;
+import com.example.graduationproject.parse.MapMain;
 
 public class MainActivity extends AppCompatActivity {
-
-    SurfaceView cameraView;
-    TextView textView;
-    CameraSource cameraSource;
-    final int RequestCameraPermissionID = 1001;
-
-    @Override
-    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
-        switch (requestCode){
-            case RequestCameraPermissionID: {
-                if (grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-                    if (ActivityCompat.checkSelfPermission(this, Manifest.permission.CAMERA) != PackageManager.PERMISSION_GRANTED) {
-                        return;
-                    }
-                    try {
-                        cameraSource.start(cameraView.getHolder());
-                    } catch (IOException e) {
-                        e.printStackTrace();
-                    }
-                }
-            }
-        }
-    }
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        cameraView = (SurfaceView)findViewById(R.id.surface_view);
-        textView = (TextView)findViewById(R.id.text_view);
+        Button loginBtn = findViewById(R.id.btnLogin);
+        Button ocrBtn = findViewById(R.id.btnOcr);
+        Button mapBtn = findViewById(R.id.btnMap);
 
-        TextRecognizer textRecognizer = new TextRecognizer.Builder(this).build();
-        if(!textRecognizer.isOperational())
-        {
-            Log.w("MainActivity","Detector dependencies are not yet available");
-        }
-        else{
-
-            cameraSource = new CameraSource.Builder(getApplicationContext(), textRecognizer)
-                    .setFacing(CameraSource.CAMERA_FACING_BACK)
-                    .setRequestedPreviewSize(1280,1024)
-                    .setRequestedFps(2.0f)
-                    .setAutoFocusEnabled(true)
-                    .build();
-            cameraView.getHolder().addCallback(new SurfaceHolder.Callback(){
-
-                @Override
-                public void surfaceCreated(SurfaceHolder holder) {
-
-                    try{
-                        if(ActivityCompat.checkSelfPermission(getApplicationContext(), Manifest.permission.CAMERA) != PackageManager.PERMISSION_GRANTED){
-                            ActivityCompat.requestPermissions(MainActivity.this,
-                                    new String[]{Manifest.permission.CAMERA},
-                                    RequestCameraPermissionID);
-                            return;
-                        }
-
-                        cameraSource.start(cameraView.getHolder());
-                    } catch (IOException e) {
-                        e.printStackTrace();
-                    }
-                }
-
-                @Override
-                public void surfaceChanged(SurfaceHolder holder, int format, int width, int height) {
-
-                }
-
-                @Override
-                public void surfaceDestroyed(SurfaceHolder holder) {
-                    cameraSource.stop();
-                }
-            });
-
-            textRecognizer.setProcessor(new Detector.Processor<TextBlock>() {
-                @Override
-                public void release() {
-
-                }
-
-                @Override
-                public void receiveDetections(Detector.Detections<TextBlock> detections) {
-                    Log.d("Main","receiveDetections");
-                    final SparseArray<TextBlock> items = detections.getDetectedItems();
-                    if(items.size() != 0){
-                        textView.post(new Runnable() {
-                            @Override
-                            public void run() {
-                                StringBuilder stringBuilder = new StringBuilder();
-                                for(int i = 0; i<items.size(); ++i){
-                                    TextBlock item = items.valueAt(i);
-                                    stringBuilder.append(item.getValue());
-                                    stringBuilder.append("\n");
-                                }
-                                textView.setText(stringBuilder.toString());
-                            }
-                        });
-                    }
-                }
-            });
-        }
+        loginBtn.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                //승환님 액티비티 연결해주세용!
+            }
+        });
+        ocrBtn.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent intent = new Intent(getApplicationContext(), com.example.graduationproject.ocr.MainActivity.class);
+                startActivity(intent);
+            }
+        });
+        mapBtn.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent intent = new Intent(getApplicationContext(), MapMain.class);
+                startActivity(intent);
+            }
+        });
     }
 }
