+++ app/src/main/java/com/example/graduationproject/parse/MapMain.java
@@ -1,0 +1,6 @@
+package com.example.graduationproject.parse;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+public class MapMain extends AppCompatActivity {
+}
